Finished get_initial_config function and moved to the database folder
